refactor(analytics): extract the shared Cursor sqlite helpers and drop the Cost fold-in
Keeps the code-review changes that stand on their own: the `cursor.sqlite.ts`
extraction the three readers now share, the parameterised `LIKE` in the usage
fetch, and the orphaned JSDoc left behind by the Team Analytics removal.

Removes `cursorCostContrib()` and its Cost-tab call sites. Folding the imported
CSV into one tab by special case left Overview and Cost disagreeing, kept 39.9M
tokens out of "Tokens by model", and vanished entirely when no local Cursor
session existed. The replacement converts the CSV into ordinary sessions so it
flows everywhere by construction.
